getMultiSigSetup now takes an "external id" as well as the public key.
An external id is a uuid for a user that is used for the sole purpose of looking up the multisig setup.
It is used instead of user id so that people do not accidentally type in someone else's user id and activate their account with another user's key.

Also activateAccount no longer requires an email, and publicKey is in the body instead of the path.